refactor code into modules, allowing for more flexibility.
